requirements.txt
transformers
transformers[sentencepiece]
sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
pyyaml
matplotlib
torch
notebook
boto3
mypy-boto3-s3
Python-box==6.0.2
ensure==1.0.4
fastapi==0.78.0
uvicorn==0.18.3
jinja2==3.1.2
datasets
accelerate
-e .
